+++ frontend/src/components/__tests__/SeqvarDetails/ClinvarCard.spec.ts
