+++ frontend/src/components/__tests__/SeqvarDetails/FreqsCard/AutosomalFreqs.spec.ts
