+++ frontend/src/components/__tests__/SeqvarDetails/FreqsCard/MitochondrialFreqs.spec.ts
