+++ frontend/src/components/__tests__/SeqvarDetails/VariantScoresCard/ScoreDisplay.spec.ts
