+++ frontend/src/components/__tests__/SeqvarDetails/VariantScoresCard/UcscConservation.spec.ts
